// api/src/container/Common/types.ts
export enum ROUTES {
  createMessage = "createMessage",
  createChat = "createChat",
  getConversations = "getConversations",
  getConversationChats = "getConversationChats",
  createComment = "createComment",
  getAllComments = "getAllComments",
  readAllNotification = "readAllNotification",
  readNotification = "readNotification",
}